feat(cogni-knowledge): extract shared wiki-grounding discovery primitive (closes #494)
Extract wiki-query's index→select→read→score discovery out of wiki-coverage.py
into one shared scripts/wiki-grounding.py primitive (rank_pages + rank CLI),
consumed by both the inverted-pipeline read-side (wiki-coverage.py, now a thin
per-sub-question caller) and the re-homed query skill. Behaviour byte-for-byte:
the existing coverage + bilingual tests pass unchanged; new test_wiki_grounding.sh
proves wiki-grounding rank == wiki-coverage score for the same sub-question.

Runs on the vendored engine via _knowledge_lib (zero installed-cogni-wiki
dependency); the vendored _wikilib.iter_pages 11-type page-walk is deferred
(FRONTMATTER_RE reconciliation risk against the bilingual cases).
