+++ cogni-knowledge/agents/source-curator.md
@@ -62,7 +62,7 @@ You **do not cobrowse**. Browser-assisted recovery of WebFetch misses is Phase 3
 | `KNOWLEDGE_ROOT` | Yes | Absolute path to the knowledge-base root (the dir containing `.cogni-knowledge/`). Forwarded to `fetch-cache.py` as `--knowledge-root` in Phase 4. |
 | `MAX_AGE_DAYS` | No | Cache freshness window in days (default 30; from `binding.curator_defaults.fetch_cache_max_age_days`). Forwarded to `fetch-cache.py fetch --max-age-days` in Phase 4. |
 | `WIKI_ROOT` | Yes | Absolute path to the bound wiki root (the dir containing `.cogni-wiki/config.json` and `wiki/`). Resolved by the orchestrator from `binding.wiki_path`. Used in Phase 1 to `Read` already-covering wiki pages for read-before-web narrowing. Same param the Phase-4 `source-ingester` already takes. |
-| `WIKI_COVERAGE_PATH` | No | Absolute path to the run's wiki-coverage manifest (`<project>/.metadata/wiki-coverage.json`), resolved **once** by the orchestrator (`knowledge-curate` Step 0.5) via `wiki-coverage.py`. Read in Phase 0; drives Phase-1 query narrowing. **Absent / unreadable ⇒ behave exactly as today** (full search) — read-before-web is an optimization, never a hard dependency. |
+| `WIKI_COVERAGE_PATH` | No | Absolute path to the run's wiki-coverage manifest (`<project>/.metadata/wiki-coverage.json`), resolved **once** by the orchestrator (`knowledge-curate` Step 0.5) via `wiki-coverage.py` — which is now a thin caller of the shared `wiki-grounding.py` discovery primitive (the single index→select→read→score mechanism the FMO ships, consumed by both this read-side and the re-homed query skill). Read in Phase 0; drives Phase-1 query narrowing. **Absent / unreadable ⇒ behave exactly as today** (full search) — read-before-web is an optimization, never a hard dependency. |
 | `CURRENT_YEAR` | No | Four-digit year. Used for recency-aware queries. |
 
 Market configuration is **not** resolved by this agent. The orchestrator (`knowledge-curate`) runs cogni-workspace's `get-market-config.py --plugin research --market <MARKET>` **once** per run — joining the canonical registry at `cogni-workspace/references/supported-markets-registry.json` with the research plugin overlay — validates it (it aborts the run if the market resolves to the `_default` fallback), and writes the merged-config envelope to `MARKET_CONFIG_PATH`. This agent just **reads that file** (Phase 0). Resolving once in skill context — where the env is consistent — removes the per-agent `WORKSPACE_PLUGIN_ROOT` glob that made one shard silently fall back to `_default` while siblings loaded the real market, and still reaches zero cogni-research code at runtime, honouring the clean-break commitment.
@@ -77,7 +77,7 @@ Phase 0 → Phase 1 → Phase 2 → Phase 3 → Phase 4
 
 1. Read `<PROJECT_PATH>/.metadata/plan.json`. Locate the sub-question with `id == SUB_QUESTION_ID`. Extract `query`, `search_guidance`, `candidate_domains[]`.
 2. Read the market config from `MARKET_CONFIG_PATH` (the orchestrator-resolved `get-market-config.py` envelope). Parse the envelope's **`data`** field — the merged market config — and store it as `market_config` (this is the shape Phases 1 and 3 consume: `market_config.authority_sources`, `market_config.local_query_tips`, …). Do **not** re-resolve the config and do **not** fall back to `_default`: the orchestrator already validated it (or aborted the run). If `MARKET_CONFIG_PATH` is missing/unreadable, is not valid JSON, or the envelope has no `data`, this is a **hard error** (defence-in-depth) — return the failure summary `{"ok": false, "sub_question_id": "<SUB_QUESTION_ID>", "reason": "market_config_unavailable"}` and stop. The orchestrator records it in `failed_curators[]`.
-3. **Load this sub-question's wiki coverage (read-before-web).** If `WIKI_COVERAGE_PATH` is provided and readable, parse the envelope's `data.sub_questions[]` and locate the entry whose `sq_id == SUB_QUESTION_ID`. Keep its `coverage_verdict` (`covered` / `partial` / `uncovered`) and `covered_pages[]` (each carrying `slug`, `type`, `page_path`, `title`, `overlap_score`). If `WIKI_COVERAGE_PATH` is absent, unreadable, not valid JSON, or has no entry for this sub-question, treat the verdict as **`uncovered`** — this is **not** an error (unlike the market config above). Read-before-web is an optimization; missing coverage just means a full search.
+3. **Load this sub-question's wiki coverage (read-before-web).** The manifest is produced by the shared `wiki-grounding.py` discovery primitive (via `wiki-coverage.py` at `knowledge-curate` Step 0.5) — the same index→select→read→score mechanism the re-homed query skill uses, so the pages surfaced here are exactly the pages a direct query would ground against. If `WIKI_COVERAGE_PATH` is provided and readable, parse the envelope's `data.sub_questions[]` and locate the entry whose `sq_id == SUB_QUESTION_ID`. Keep its `coverage_verdict` (`covered` / `partial` / `uncovered`) and `covered_pages[]` (each carrying `slug`, `type`, `page_path`, `title`, `overlap_score`). If `WIKI_COVERAGE_PATH` is absent, unreadable, not valid JSON, or has no entry for this sub-question, treat the verdict as **`uncovered`** — this is **not** an error (unlike the market config above). Read-before-web is an optimization; missing coverage just means a full search.
 4. Confirm `BATCH_OUTPUT_PATH`'s parent directory exists; create if not.
 
 ### Phase 1: Search Query Generation
